% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kar.r
\name{f3}
\alias{f3}
\title{Persian Text Normalization and Stemming}
\usage{
f3(x)
}
\arguments{
\item{x}{A character vector of Persian text.}
}
\value{
A character vector of normalized and stemmed text.
}
\description{
This function normalizes Persian text by replacing specific characters
and applies stemming.
}
\examples{
text <- c("\u067e\u0627\u0631\u0633\u06cc \u0645\u062a\u0646")
normalized_text <- f3(text)
}
