% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kar.r
\name{fun.all.sums}
\alias{fun.all.sums}
\title{Apply Suffix Modifications to Persian Words}
\usage{
fun.all.sums(v, TYPE)
}
\arguments{
\item{v}{A character vector of Persian words.}

\item{TYPE}{A vector of suffix types for modification.}
}
\value{
A modified character vector.
}
\description{
This function iteratively applies a series of suffix modifications to a vector of Persian words.
}
\examples{
words <- c("\u0632\u06cc\u0628\u0627\u06cc\u06cc\u0645\u0627\u0646")
modified_words <- fun.all.sums(words, TYPE)
}
