% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kar.r
\name{fungan}
\alias{fungan}
\title{Persian Suffix Modification for '\u06af\u0627\u0646' Suffix}
\usage{
fungan(v)
}
\arguments{
\item{v}{A character vector of Persian words.}
}
\value{
A modified character vector.
}
\description{
This function modifies Persian words ending with '\u06af\u0627\u0646' suffix.
}
\examples{
words <- c("\u062f\u0648\u0633\u062a\u0627\u0646")
modified_words <- fungan(words)
}
