% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kar.r
\name{fungi}
\alias{fungi}
\title{Persian Suffix Modification}
\usage{
fungi(v)
}
\arguments{
\item{v}{A character vector of Persian words.}
}
\value{
A modified character vector.
}
\description{
This function modifies Persian words ending with '\u06af\u06cc' suffix.
}
\examples{
words <- c("\u0632\u06cc\u0628\u0627\u06cc\u06cc")
modified_words <- fungi(words)
}
