% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kar.r
\name{funmi}
\alias{funmi}
\title{Modify Persian Words Starting with '\u0645\u06cc'}
\usage{
funmi(v)
}
\arguments{
\item{v}{A character vector of Persian words.}
}
\value{
A modified character vector.
}
\description{
This function modifies Persian words starting with the prefix '\u0645\u06cc'.
}
\examples{
words <- c("\u0645\u06cc\u0631\u0648\u062f")
modified_words <- funmi(words)
}
