% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kar.r
\name{LEMMA}
\alias{LEMMA}
\title{Persian Lemmatization}
\usage{
LEMMA(Y, TYPE)
}
\arguments{
\item{Y}{A character vector of Persian words.}

\item{TYPE}{A vector of suffix types for modification.}
}
\value{
A vector of lemmatized Persian words.
}
\description{
This function performs lemmatization on a vector of Persian words.
}
\examples{
words <- c("\u062f\u0648\u0633\u062a\u0627\u0646\u0645")
lemmatized_words <- LEMMA(words, TYPE)
}
