% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kar.r
\docType{data}
\name{TYPE}
\alias{TYPE}
\title{Persian Suffixes}
\format{
An object of class \code{character} of length 39.
}
\usage{
TYPE
}
\description{
A vector of common Persian suffixes used for text processing.
}
\keyword{datasets}
